+++ src/taoensso/timbre/profiling.clj
@@ -207,32 +207,56 @@
 
 ;;;;
 
-(defmacro defnp "Like `defn` but wraps fn bodies with `p` macro."
-  {:arglists
-   '([name doc-string? attr-map? [params*] prepost-map? body]
-     [name doc-string? attr-map? ([params*] prepost-map? body)+ attr-map?])}
-  [name' & sigs]
-  (let [[name' sigs]  (enc/name-with-attrs name' sigs)
-        single-arity? (vector? (first sigs))
-        [sigs func->str]
-        (if single-arity?
-          [(list sigs) (fn [name' _params] (name name'))]
-          [sigs        (fn [name'  params] (str (name name') \_ (count params)))])
-
+(defn fn-sigs "Implementation detail."
+  [fn-name sigs]
+  (let [single-arity? (vector? (first sigs))
+        sigs    (if single-arity? (list sigs) sigs)
+        get-pid (if single-arity?
+                  (fn [fn-name _params]      (name fn-name))
+                  (fn [fn-name  params] (str (name fn-name) \_ (count params))))
         new-sigs
-        (map (fn [[params & others]]
-               (let [has-prepost-map? (and (map? (first others)) (next others))
-                     [prepost-map & body]
-                     (if has-prepost-map?
-                       others
-                       (cons {} others))]
-                 `(~params ~prepost-map (pspy ~(func->str name' params) ~@body))))
+        (map
+          (fn [[params & others]]
+            (let [has-prepost-map?      (and (map? (first others)) (next others))
+                  [?prepost-map & body] (if has-prepost-map? others (cons nil others))]
+              (if ?prepost-map
+                `(~params ~?prepost-map (pspy ~(get-pid fn-name params) ~@body))
+                `(~params               (pspy ~(get-pid fn-name params) ~@body)))))
           sigs)]
-    `(defn ~name' ~@new-sigs)))
+    new-sigs))
+
+(defmacro fnp "Like `fn` but wraps fn bodies with `p` macro."
+  {:arglists '([name?  [params*] prepost-map? body]
+               [name? ([params*] prepost-map? body)+])}
+  [& sigs]
+  (let [[?fn-name sigs] (if (symbol? (first sigs))
+                          [(first sigs) (next sigs)]
+                          [nil sigs])
+        new-sigs (fn-sigs (or ?fn-name 'anonymous-fn) sigs)]
+    (if ?fn-name
+      `(fn ~?fn-name ~@new-sigs)
+      `(fn           ~@new-sigs))))
 
 (comment
-  (defnp foo "Docstring "[x] "boo" (* x x))
-  (macroexpand '(defnp foo "Docstring" [x] "boo" (* x x)))
+  (fn-sigs "foo" '([x] (* x x)))
+  (macroexpand '(fnp [x] (* x x)))
+  (macroexpand '(fn   [x] (* x x)))
+  (macroexpand '(fnp [x] {:pre [x]} (* x x)))
+  (macroexpand '(fn   [x] {:pre [x]} (* x x))))
+
+(defmacro defnp "Like `defn` but wraps fn bodies with `p` macro."
+  {:arglists
+   '([name doc-string? attr-map?  [params*] prepost-map? body]
+     [name doc-string? attr-map? ([params*] prepost-map? body)+ attr-map?])}
+  [& sigs]
+  (let [[fn-name sigs] (enc/name-with-attrs (first sigs) (next sigs))
+        new-sigs       (fn-sigs fn-name sigs)]
+    `(defn ~fn-name ~@new-sigs)))
+
+(comment
+  (defnp foo "Docstring "[x] (* x x))
+  (macroexpand '(defnp foo "Docstring" [x] (* x x)))
+  (macroexpand '(defn  foo "Docstring" [x] (* x x)))
   (macroexpand '(defnp foo "Docstring" ([x]   (* x x))
                                        ([x y] (* x y))))
   (profile :info :defnp-test (foo 5)))